refactor: add LoggingModule func for joins
